DBZ-6209 Extract additional common code
